+++ presentation/src/main/java/com/tamerbarsbay/depothouston/presentation/presenter/ArrivalListPresenter.java
@@ -19,9 +19,6 @@ import java.util.List;
 import javax.inject.Inject;
 import javax.inject.Named;
 
-/**
- * Created by Tamer on 7/24/2015.
- */
 @PerActivity
 public class ArrivalListPresenter extends DefaultSubscriber<List<Arrival>> implements Presenter {
