+++ presentation/src/main/java/com/tamerbarsbay/depothouston/presentation/presenter/BuildItineraryPresenter.java
@@ -4,9 +4,6 @@ import com.tamerbarsbay.depothouston.domain.Itinerary;
 import com.tamerbarsbay.depothouston.domain.interactor.DefaultSubscriber;
 import com.tamerbarsbay.depothouston.presentation.internal.di.PerActivity;
 
-/**
- * Created by Tamer on 7/24/2015.
- */
 @PerActivity
 public class BuildItineraryPresenter extends DefaultSubscriber<Itinerary> implements Presenter {
