+++ presentation/src/main/java/com/tamerbarsbay/depothouston/presentation/presenter/IncidentListPresenter.java
@@ -19,9 +19,6 @@ import java.util.List;
 import javax.inject.Inject;
 import javax.inject.Named;
 
-/**
- * Created by Tamer on 7/24/2015.
- */
 @PerActivity
 public class IncidentListPresenter extends DefaultSubscriber<List<Incident>> implements Presenter {
