+++ presentation/src/main/java/com/tamerbarsbay/depothouston/presentation/presenter/MapSearchPresenter.java
@@ -1,0 +1,7 @@
+package com.tamerbarsbay.depothouston.presentation.presenter;
+
+import com.tamerbarsbay.depothouston.presentation.internal.di.PerActivity;
+
+@PerActivity
+public class MapSearchPresenter {
+}
