+++ presentation/src/main/res/layout/activity_maps.xml
@@ -3,5 +3,5 @@
           android:layout_width="match_parent"
           android:layout_height="match_parent"
           android:id="@+id/map"
-          tools:context="com.tamerbarsbay.depothouston.presentation.view.activity.MapsActivity"
+          tools:context="com.tamerbarsbay.depothouston.presentation.view.activity.MapSearchActivity"
           android:name="com.google.android.gms.maps.SupportMapFragment"/>
